fix: bound full-v1 verifier enumeration
