+++ tests/run-tests.js
@@ -14,10 +14,6 @@ const path = require('path');
 
 process.env.NODE_ENV = process.env.NODE_ENV || 'test';
 
-// Unit tests must never inherit operator credentials or model overrides.
-const testHome = fs.mkdtempSync(path.join(os.tmpdir(), 'zeroshot-unit-home-'));
-const testSettingsFile = path.join(testHome, '.zeroshot', 'settings.json');
-
 const testFileAliases = new Map([
   ['provider-cli-builder', ['tests/provider-cli-builder.test.js']],
   ['providers', ['tests/providers/**/*.test.js']],
@@ -39,12 +35,23 @@ const defaultTestFiles = [
 ];
 const hasExplicitTestFile = args.some((arg) => !arg.startsWith('-'));
 const mochaArgs = hasExplicitTestFile ? args : [...defaultTestFiles, ...args];
+// Unit tests must never inherit operator credentials or model overrides.
+const testHome = fs.mkdtempSync(path.join(os.tmpdir(), 'zeroshot-test-home-'));
+const testSettingsFile = path.join(testHome, '.zeroshot', 'settings.json');
+const testEnvironment = {};
+for (const [key, value] of Object.entries(process.env)) {
+  if (key.startsWith('ZEROSHOT_') || key.startsWith('CMDPROOF_')) continue;
+  testEnvironment[key] = value;
+}
+fs.mkdirSync(path.dirname(testSettingsFile), { recursive: true });
+fs.writeFileSync(testSettingsFile, '{}\n', 'utf8');
 const child = spawn('npx', ['mocha', '--parallel', ...mochaArgs], {
   stdio: ['inherit', 'pipe', 'pipe'],
   env: {
-    ...process.env,
+    ...testEnvironment,
     HOME: testHome,
     USERPROFILE: testHome,
+    ZEROSHOT_HOME: testHome,
     ZEROSHOT_SETTINGS_FILE: testSettingsFile,
   },
 });
